// api/src/game/building/configuration/buildingType.ts
export enum BuildingType {
  CRYSTAL_MINE = 'CRYSTAL_MINE',
  DEUTERIUM_SYNTHESIZER = 'DEUTERIUM_SYNTHESIZER',
  METAL_MINE = 'METAL_MINE',
  FUSION_REACTOR = 'FUSION_REACTOR',
  SOLAR_PLANT = 'SOLAR_PLANT',
  SOLAR_SATELLITE = 'SOLAR_SATELLITE',
  CRYSTAL_STORAGE = 'CRYSTAL_STORAGE',
  METAL_STORAGE = 'METAL_STORAGE',
  DEUTERIUM_TANK = 'DEUTERIUM_TANK',
  ROBOTICS_FACTORY = 'ROBOTICS_FACTORY',
  SHIPYARD = 'SHIPYARD',
  RESEARCH_LAB = 'RESEARCH_LAB',
  ALLIANCE_DEPOT = 'ALLIANCE_DEPOT',
  MISSILE_SILO = 'MISSILE_SILO',
  NANITE_FACTORY = 'NANITE_FACTORY',
  TERRAFORMER = 'TERRAFORMER',
}